+++ index.ts
@@ -111,13 +111,24 @@ client.on('message', async (msg: Message & { channel: { name: string } }) => {
             });
         });
     } else if (parsed.order === '!説明') {
-        const info_text = '★忙しすぎるあなたに代わってチャンネルを作成します。\n\n** ● 募集を立てたいとき ● **\n サーバー内のいずれかのテキストチャンネル内で`!募集 チャンネル名`と発言してください。「学園掲示板A」カテゴリ内に新規テキストチャンネルが作成されます。\n' +
-            '　例）`!募集　1224 伝説の入り口(ARA2E)`\n\n' +
-            '** ● 教室を立てたいとき ● **\n `!教室　教室名`と発言してください。\n「教室棟」「教室棟VC」カテゴリにそれぞれチャンネルが作成されます。\n' +
-            '　例）`!教室 伝説の入り口（ARA2E)`\n\n' +
-            '** ● キャンペーン用の教室を立てたいとき ● **\n `!キャンペーン　教室名`と発言してください。\n「CP用教室棟」「CP用教室棟VC」カテゴリにそれぞれチャンネルが作成されます。\n' +
-            '　例）`!キャンペーン ファーストクエスト（ARA2E)`\n\n' +
-            '** ● チャンネルの削除を行いたいとき ● **\n 上記手順で作成されたテキストチャンネル内で、チャンネル作成を行ったユーザーが`!削除`と発言してください。\n\n' +
+        const info_text = '★忙しすぎるあなたに代わってチャンネルを作成します。\n\n' +
+            '** ● 募集を立てたいとき ● **\n' +
+            '> サーバー内のいずれかのテキストチャンネル内で`!募集 チャンネル名`と発言してください。\n' +
+            '> 「学園掲示板A」カテゴリ内に新規テキストチャンネルが作成されます。\n' +
+            '> 例）`!募集　1224 伝説の入り口(ARA2E)`\n\n' +
+            '** ● 教室を立てたいとき ● **\n `!教室　教室名`と発言してください。\n' +
+            '> 「教室棟」「教室棟VC」カテゴリにそれぞれチャンネルが作成されます。\n' +
+            '> 例）`!教室 伝説の入り口（ARA2E)`\n\n' +
+            '** ● キャンペーン用の教室を立てたいとき ● **\n' +
+            '> `!キャンペーン　教室名`と発言してください。\n' +
+            '> 「CP用教室棟」「CP用教室棟VC」カテゴリにそれぞれチャンネルが作成されます。\n' +
+            '> 例）`!キャンペーン ファーストクエスト（ARA2E)`\n\n' +
+            '** ● チャンネルの名前変更を行いたいとき ● **\n' +
+            '> 作成されたテキストチャンネル内で、チャンネル作成を行ったユーザーが`!変更 新しい教室名`と発言してください。\n' +
+            '> 例）`!変更 〆1224伝説の入り口`\n\n' +
+            '** ● チャンネルの削除を行いたいとき ● **\n' +
+            '> 作成されたテキストチャンネル内で、チャンネル作成を行ったユーザーが`!削除`と発言してください。\n\n' +
+            '※一つのチャンネルに対する各種操作は、一定時間内に実行可能な回数に制限があります。連続で命令を行うと、最大10分後に反映されたりすることがありますのでご了承ください。\n' +
             '※チャンネルの名前については、学園のルールに準拠するようにしてください。';
         msg.channel.send(info_text)
     } else if (parsed.order === '!教室' || parsed.order === '!キャンペーン') {    // チャンネルを作成する
@@ -186,17 +197,19 @@ client.on('message', async (msg: Message & { channel: { name: string } }) => {
                     if (tc) {
                         // @ts-ignore
                         tc.setName(new_title, 'reason: test').then((_tc) => {
-                            client.channels.fetch(channels[i].voice_channel, false, true).then(vc => {
-                                if (vc) {
-                                    // @ts-ignore
-                                    vc.setName(new_title).then(() => {
+                            if (channels[i].voice_channel) {
+                                client.channels.fetch(channels[i].voice_channel, false, true).then(vc => {
+                                    if (vc) {
                                         // @ts-ignore
-                                        channels[i].update({channel_name: new_title}).then();
-                                    });
-                                } else {
-                                    msg.channel.send(`チャンネル名を「${new_title}」に変更しました。`);
-                                }
-                            });
+                                        vc.setName(new_title).then(() => {
+                                            // @ts-ignore
+                                            channels[i].update({channel_name: new_title}).then();
+                                        });
+                                    } else {
+                                        msg.channel.send(`チャンネル名を「${new_title}」に変更しました。`);
+                                    }
+                                });
+                            }
                         }).catch(console.error);
                     }
                 });
@@ -214,14 +227,16 @@ client.on('message', async (msg: Message & { channel: { name: string } }) => {
                 client.channels.fetch(channels[i].text_channel, false, true).then(tc => {
                     if (tc) {
                         tc.delete().then((tc_deleted: Channel) => {
-                            client.channels.fetch(channels[i].voice_channel, false, true).then(vc => {
-                                if (vc) {
-                                    vc.delete().then((vc_deleted: Channel) => {
-                                        // @ts-ignore
-                                        channels[i].update({is_deleted: true}).then();
-                                    });
-                                }
-                            }).catch(console.error);
+                            if (channels[i].voice_channel) {
+                                client.channels.fetch(channels[i].voice_channel, false, true).then(vc => {
+                                    if (vc) {
+                                        vc.delete().then((vc_deleted: Channel) => {
+                                            // @ts-ignore
+                                            channels[i].update({is_deleted: true}).then();
+                                        });
+                                    }
+                                }).catch(console.error);
+                            }
                         });
                     }
                 }).catch(console.error);
